feat: change definitions API to Datamuse
The old one did not have definitions for most of the  related words,
synonyms or antonyms provided by the word finder API, which is Datamuse, so it had to be
changed to provide a consistent experience. Also Datamuse is much easier to
work with, it offers more and better organized data.